New version of encrypt
Allows customizing the RSA and data encryption key passed into the
encrypt function. Passing the data encryption key allows the user more
control over what happens with the dek. The other component here is that
the dek is cached by default. If encrypting multiple items it will reuse
the same dek for efficiency reasons. Encrypting a bunch of things with
RSA can be slow.